% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMobile.R
\name{add_f7_tooltip}
\alias{add_f7_tooltip}
\title{Create a tooltip on the server side}
\usage{
add_f7_tooltip(
  id = NULL,
  selector = NULL,
  options,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Target id.}

\item{selector}{Target selector.}

\item{options}{Tooltip options}

\item{session}{Shiny session object}
}
\description{
Create a tooltip on the server side
}
