% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSUICode-dependencies.R
\name{add_OSUICode_deps}
\alias{add_OSUICode_deps}
\title{OSUICode dependencies utils}
\usage{
add_OSUICode_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches OSUICode dependencies to the given tag
}
