% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{add_tabler_deps}
\alias{add_tabler_deps}
\title{Create Tabler dependencies}
\usage{
add_tabler_deps(tag)
}
\arguments{
\item{tag}{Tag on which to add dependencies. We usually target the body.}
}
\description{
Add all necessary dependencies so that Tabler works
}
\seealso{
\link{tabler_page}.
}
