% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes.R
\name{box}
\alias{box}
\title{Create a custom shinydashboard box with interactive capabilities}
\usage{
box(
  ...,
  id = NULL,
  title = NULL,
  footer = NULL,
  background = NULL,
  width = 6,
  height = NULL,
  collapsible = FALSE,
  collapsed = FALSE
)
}
\arguments{
\item{...}{Contents of the box.}

\item{id}{Box unique id. \link{updateBox2} target.}

\item{title}{Optional title.}

\item{footer}{Optional footer text.}

\item{background}{If NULL (the default), the background of the box will be
white. Otherwise, a color string. Valid colors are listed in
\link[shinydashboard]{validColors}.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default valueBox width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{height}{The height of a box, in pixels or other CSS unit. By default
the height scales automatically with the content.}

\item{collapsible}{If TRUE, display a button in the upper right that allows
the user to collapse the box.}

\item{collapsed}{If TRUE, start collapsed. This must be used with
\code{collapsible=TRUE}.}
}
\description{
Create a custom shinydashboard box with interactive capabilities
}
