% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bslib-utils.R
\name{bs_accordion}
\alias{bs_accordion}
\title{Bootstrap accordion unified wrapper}
\usage{
bs_accordion(id, items)
}
\arguments{
\item{id}{Accordion unique id.}

\item{items}{Slot for \link{bs_accordion_item}.}
}
\description{
Provide support for BS4 and BS5
}
