% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bslib-utils.R
\name{bs_accordion_item}
\alias{bs_accordion_item}
\title{Bootstrap unified accordion item wrapper}
\usage{
bs_accordion_item(title, content, active = FALSE)
}
\arguments{
\item{title}{Item title.}

\item{content}{Item content.}

\item{active}{Whether to open the itm at start. Default to FALSE.}
}
\description{
Provide support for BS4 and BS5
}
