% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bslib-utils.R
\name{bs_badge}
\alias{bs_badge}
\title{Unified Bootstrap badge}
\usage{
bs_badge(text, color = NULL)
}
\arguments{
\item{text}{Badge text.}

\item{color}{Badge color.}
}
\value{
A Bootstrap badge with different class depending on
the currently used Bootstrap theme.
}
\description{
Badge providing support for BS3, BS4 and BS5
}
