% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bslib-utils.R
\docType{data}
\name{bslib_dark_theme}
\alias{bslib_dark_theme}
\title{Dark theme for bslib demo apps}
\format{
An object of class \code{bs_version_4} (inherits from \code{bs_theme}, \code{sass_bundle}) of length 1.
}
\usage{
bslib_dark_theme
}
\description{
Dark theme for bslib demo apps
}
\keyword{datasets}
