% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming-with-fresh.R
\name{customize_bs4Dash}
\alias{customize_bs4Dash}
\title{Customize bs4Dash with fresh}
\usage{
customize_bs4Dash(theme)
}
\arguments{
\item{theme}{Theme built with create_theme from fresh}
}
\value{
A shiny app example
}
\description{
Customize bs4Dash with fresh
}
\examples{
if (interactive()) {
 library(shiny)
 library(bs4Dash)
 library(fresh)

 dark_theme <- create_theme(
  bs4dash_vars(
    navbar_dark_color = "#bec5cb",
    navbar_dark_active_color = "#FFF",
    navbar_dark_hover_color = "#FFF"
  ),
  bs4dash_yiq(contrasted_threshold = 10, text_dark = "#FFF", text_light = "#272c30"),
  bs4dash_layout(main_bg = "#353c42"),
  bs4dash_sidebar_dark(
    bg = "#272c30", color = "#bec5cb", hover_color = "#FFF",
    submenu_bg = "#272c30", submenu_color = "#FFF", submenu_hover_color = "#FFF"
  ),
  bs4dash_status(dark = "#272c30"),
  bs4dash_color(gray_900 = "#FFF", white = "#272c30")
 )

 customize_bs4Dash(dark_theme)

}
}
