% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming-with-fresh.R
\name{customize_shinydashboard}
\alias{customize_shinydashboard}
\title{Customize shinydashboard with fresh}
\usage{
customize_shinydashboard(theme)
}
\arguments{
\item{theme}{Theme built with create_theme from fresh}
}
\value{
A shiny app example
}
\description{
Customize shinydashboard with fresh
}
