% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{customTextInput}
\alias{customTextInput}
\title{Create a custom  Shiny text input}
\usage{
customTextInput(
  inputId,
  label,
  value = "",
  width = NULL,
  placeholder = NULL,
  binding_step
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input, e.g. '400px', or '100\%'; see \link[shiny]{validateCssUnit}.}

\item{placeholder}{A character string giving the user a
hint as to what can be entered into the control.
Internet Explorer 8 and 9 do not support this option.}

\item{binding_step}{This parameter is only for the book purpose. This is to be able
to follow each step of the input binding creation of Chapter 5.4. When you develop
your own template, you don't need this. Provides the fully working binding instead!}
}
\value{
A text input control that can be added to a UI definition.
}
\description{
This is to showcase how to develop Shiny custom inputs (see Chapter 5.4)
}
\examples{
if (interactive()) {
library(shiny)

 # This example corresponds to section 5.4.2.1
 customTextInputExample(1)

}
}
