% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{customTextInputDeps}
\alias{customTextInputDeps}
\title{Add necessary dependencies for the \link{customTextInput}}
\usage{
customTextInputDeps(step)
}
\arguments{
\item{step}{Input binding step. Each step corresponds to a specific method
of the full input binding. For instance, step 1 only includes the find method,
step 2 contains, find + getValue...}
}
\value{
An html dependency for \link{customTextInput}.
}
\description{
The step provides a way to follow the Chapter 1 book section.
For instance, if step = 1, it means that we consider the 5.4.2.1 Find the input section, and so on,
until the fully complete binding.
}
