% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{customTextInputExample}
\alias{customTextInputExample}
\title{Wrapper for Shiny App example}
\usage{
customTextInputExample(binding_step, value = "Data Summary")
}
\arguments{
\item{binding_step}{Input binding step. See \link{customTextInput}.}
}
\value{
A Shiny App example
}
\description{
Wrapper for Shiny App example
}
