% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{customTextInputExampleBis}
\alias{customTextInputExampleBis}
\title{Wrapper for Shiny App example}
\usage{
customTextInputExampleBis()
}
\value{
A Shiny App example
}
\description{
Wrapper for Shiny App example
}
