% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{customTextInputHandlerExample}
\alias{customTextInputHandlerExample}
\title{Wrapper for Shiny App example with input handler}
\usage{
customTextInputHandlerExample(binding_step, value = "Data Summary")
}
\arguments{
\item{binding_step}{Input binding step. See \link{customTextInput}.}

\item{value}{Default value. Dates are handled with a custom input handler.}
}
\value{
A Shiny App example
}
\description{
Wrapper for Shiny App example with input handler
}
