% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-handlers.R
\docType{data}
\name{dashboardUserImage}
\alias{dashboardUserImage}
\title{User image from dropdownMenu items}
\format{
An object of class \code{character} of length 1.
}
\usage{
dashboardUserImage
}
\description{
User image from dropdownMenu items
}
\keyword{datasets}
