% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDropdownItem.R
\name{dropdownMenu}
\alias{dropdownMenu}
\title{Create a Boostrap 4 dashboard dropdown menu}
\usage{
dropdownMenu(
  ...,
  show = FALSE,
  labelText = NULL,
  src = NULL,
  status = c("primary", "warning", "danger", "info", "success"),
  menuIcon = "bell",
  align = "right"
)
}
\arguments{
\item{...}{Slot for \link[bs4Dash]{dropdownMenuItem}.}

\item{show}{Whether to start with the dropdown open. FALSE by default.}

\item{labelText}{Dropdown label text.}

\item{src}{Dropdown link to an external ressource.}

\item{status}{Dropdown menu status. "primary", "success", "warning", "danger" or "info".}

\item{menuIcon}{Fontawesome icon (default = "bell")}

\item{align}{Menu alignment (default = "right")}
}
\description{
Build an adminLTE3 dashboard dropdown menu
}
