% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{editBindingDeps}
\alias{editBindingDeps}
\title{Create HTML dependency for edit binding example}
\usage{
editBindingDeps()
}
\value{
An HTML dependency that may be attached to a tag.
}
\description{
Create HTML dependency for edit binding example
}
