% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{editBindingExample}
\alias{editBindingExample}
\title{Example showing how to edit an input binding}
\usage{
editBindingExample()
}
\value{
A shiny app example
}
\description{
Example showing how to edit an input binding
}
