% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMobile.R
\name{f7_gauge}
\alias{f7_gauge}
\title{Gauge widget}
\usage{
f7_gauge(id, value, options = NULL)
}
\arguments{
\item{id}{Gauge unique id. Needed by \link{update_f7_instance}.}

\item{value}{Gauge value.}

\item{options}{Gauge options. Pass a list.}
}
\description{
Gauge widget
}
\examples{
if (interactive()) {
 shiny::shinyAppDir(system.file("shinyMobile-pwa", package = "OSUICode"))
}
}
