% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMobile.R
\name{f7_notif}
\alias{f7_notif}
\title{Notification widget}
\usage{
f7_notif(
  id = NULL,
  text,
  options = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Notification unique id. Needed by \link{update_f7_instance}.}

\item{text}{Notification text.}

\item{options}{List of options.}

\item{session}{Shiny session object.}
}
\description{
Notification widget
}
\examples{
if (interactive()) {
 shiny::shinyAppDir(system.file("shinyMobile/notification", package = "OSUICode"))
}
}
