% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMobile.R
\name{f7_page}
\alias{f7_page}
\title{Mobile page wrapper}
\usage{
f7_page(..., navbar, toolbar, title = NULL, options = NULL, allowPWA = TRUE)
}
\arguments{
\item{...}{Body elements}

\item{navbar}{Slot for \link{f7_navbar}.}

\item{toolbar}{Slot for \link{f7_toolbar}.}

\item{title}{Tab title.}

\item{options}{Options to configure the template.}

\item{allowPWA}{Whether to allow PWA. Defaults to TRUE.}
}
\description{
Mobile page wrapper
}
\examples{
if (interactive()) {
 shiny::shinyAppDir(system.file("shinyMobile-pwa", package = "OSUICode"))
}
}
