% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{findAttribute}
\alias{findAttribute}
\title{Find the child of the targeted element that has a specific attribute and value.}
\usage{
findAttribute(x, attr, val)
}
\arguments{
\item{x}{Parent tag.}

\item{attr}{Atttribute like id, class, data-toggle, ...}

\item{val}{Attribute value.}
}
\value{
TRUE or FALSE, depending on the search result
}
\description{
This function takes a DOM element/tag object and reccurs within it until
}
