% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket.R
\name{httpuv_app}
\alias{httpuv_app}
\title{Create demo httpuv app}
\usage{
httpuv_app(delay = NULL)
}
\arguments{
\item{delay}{To simulate computationally intense task.}
}
\value{
An httpuv powered app
}
\description{
This app is composed of an HTTP response containing a
an HTML page with a range slider and a client websocket connection.
The app has also a server websocket handler.
At the end of the day, don't forget to stop the app!
}
\examples{
if (interactive()) {
 library(OSUICode)
 my_app <- httpuv_app()
 my_app$stop()
}
}
