% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{insert_tabler_tab}
\alias{insert_tabler_tab}
\title{Insert a tab in a \link{tabler_navbar}}
\usage{
insert_tabler_tab(
  inputId,
  tab,
  target,
  position = c("before", "after"),
  select = FALSE,
  handler_type,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{Tabler navbar menu unique id.}

\item{tab}{Tab to insert. Must be a \link{tabler_tab_item}.}

\item{target}{Target tab after or before which the new tab will be inserted.}

\item{position}{Insert position: "before" or "after".}

\item{select}{Whether to select the new tab at start. Default to FALSE.}

\item{handler_type}{Only for the Book purpose. Allow to select a given
custom message handler, to produce different behavior. You don't need
this in production!}

\item{session}{Shiny session.}
}
\description{
Insert a tab in a \link{tabler_navbar}
}
\examples{
if (interactive()) {
 library(shiny)
 # See book section 15.4 (Tab events)
 # insert tab with wrong custom handler. The tab is inserted
 # but nothing is displayed inside.
 insert_tabler_tab_example(1)

 # insert tab with good custom handler.
 insert_tabler_tab_example(2)
}
}
