% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{insert_tabler_tab_2}
\alias{insert_tabler_tab_2}
\title{Insert a tab in a \link{tabler_navbar}}
\usage{
insert_tabler_tab_2(
  inputId,
  tab,
  target,
  position = c("before", "after"),
  select = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{Tabler navbar menu unique id.}

\item{tab}{Tab to insert. Must be a \link{tabler_tab_item}.}

\item{target}{Target tab after or before which the new tab will be inserted.}

\item{position}{Insert position: "before" or "after".}

\item{select}{Whether to select the new tab at start. Default to FALSE.}

\item{session}{Shiny session.}
}
\description{
This function is the right one to use. It correctly
processes all dependencies at run time.
}
