% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{insert_tabler_tab_example}
\alias{insert_tabler_tab_example}
\title{Create an example for \link{insert_tabler_tab}.}
\usage{
insert_tabler_tab_example(handler_type)
}
\arguments{
\item{handler_type}{Only for the Book purpose. Allow to select a given
custom message handler, to produce different behavior. You don't need
this in production!}
}
\value{
A shiny app.
}
\description{
Create an example for \link{insert_tabler_tab}.
}
\seealso{
\link{insert_tabler_tab}.
}
