% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDropdownItem.R
\name{insertDropdownItem}
\alias{insertDropdownItem}
\title{Insert Bootstrap 4 dropdown menu item}
\usage{
insertDropdownItem(item, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{item}{Item to insert.}

\item{session}{Shiny session object.}
}
\description{
Insert Bootstrap 4 dropdown menu item
}
\examples{
if (interactive()) {
 library(shiny)
 shinyAppDir(system.file("chapter6/add-dropdown-item", package = "OSUICode"))
}
}
