% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMessageItem.R
\name{insertMessageItem}
\alias{insertMessageItem}
\title{Insert Bootstrap 4 dropdown menu item}
\usage{
insertMessageItem(item, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{item}{Item to insert.}

\item{session}{Shiny session object.}
}
\description{
Insert Bootstrap 4 dropdown menu item
}
\examples{
if (interactive()) {
 library(shiny)
 shinyAppDir(system.file("custom-handlers/add-message-item", package = "OSUICode"))
}
}
