% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmtools-deps.R
\name{my_card}
\alias{my_card}
\title{Simple BS4 card without dependencies}
\usage{
my_card(...)
}
\arguments{
\item{...}{Card content}
}
\value{
An HTML card
}
\description{
This card is missing dependencies
}
