% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmtools-deps.R
\name{my_card_with_deps}
\alias{my_card_with_deps}
\title{Simple Material Bootstrap card with dependencies}
\usage{
my_card_with_deps(...)
}
\arguments{
\item{...}{Card content}
}
\value{
An HTML card
}
\description{
Simple Material Bootstrap card with dependencies
}
