% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmtools-deps.R
\name{my_dashboard_box}
\alias{my_dashboard_box}
\title{Standalone shinydashboard box}
\usage{
my_dashboard_box(title, status)
}
\arguments{
\item{title}{Box title.}

\item{status}{Box color.}
}
\value{
A box HTML tag.
}
\description{
box from shinydashboard you can use inside
a basic shiny app.
}
