% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bslib-utils.R
\docType{data}
\name{nes_theme}
\alias{nes_theme}
\title{NES theme for shiny}
\format{
An object of class \code{bs_version_4} (inherits from \code{bs_theme}, \code{sass_bundle}) of length 1.
}
\usage{
nes_theme
}
\description{
See https://rstudio.github.io/bslib/articles/bslib.html#add-rules
}
\keyword{datasets}
