% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemons.R
\name{runPokemonExample}
\alias{runPokemonExample}
\title{Shiny app example showing random pokemon image}
\usage{
runPokemonExample()
}
\value{
A shiny app
}
\description{
Shiny app example showing random pokemon image
}
