% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/say_hello.R
\name{say_hello_to_js}
\alias{say_hello_to_js}
\title{Function demonstrating the use of Shiny custom message handlers}
\usage{
say_hello_to_js(text, session = getDefaultReactiveDomain())
}
\arguments{
\item{text}{Message to send.}

\item{session}{Shiny session object.}
}
\description{
Send a message from R to JavaScript every 5 seconds.
The JS answer is an alert containing the message sent by R.
}
\examples{
if (interactive()) {
 library(shiny)
 shinyAppDir(system.file("chapter6/say_hello", package = "OSUICode"))
}
}
