% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMobile.R
\docType{data}
\name{shinyMobile_options}
\alias{shinyMobile_options}
\title{Default shinyMobile template options}
\format{
An object of class \code{list} of length 8.
}
\usage{
shinyMobile_options
}
\description{
Default shinyMobile template options
}
\keyword{datasets}
