% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{show_tabler_dropdown}
\alias{show_tabler_dropdown}
\title{Show a \link{tabler_dropdown} on the client}
\usage{
show_tabler_dropdown(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{Dropdown id.}

\item{session}{Shiny session}
}
\description{
Show a \link{tabler_dropdown} on the client
}
\examples{
if (interactive()) {
 ui <- tabler_page(
  tabler_navbar(
    brand_url = "https://preview-dev.tabler.io",
    brand_image = "https://preview-dev.tabler.io/static/logo.svg",
    nav_menu = NULL,
    tabler_dropdown(
      id = "mydropdown",
      title = "Dropdown",
      subtitle = "click me",
      tabler_dropdown_item(
        id = "item1",
        "Show Notification"
      ),
      tabler_dropdown_item(
        "Do nothing"
      )
    )
  ),
  tabler_body(
    tabler_button("show", "Open dropdown", width = "25\%"),
    footer = tabler_footer(
      left = "Rstats, 2020",
      right = a(href = "https://www.google.com")
    )
  )
 )
 server <- function(input, output, session) {

   observeEvent(input$show, {
     show_tabler_dropdown("mydropdown")
   })

   observeEvent(input$item1, {
     showNotification(
       "Success",
       type = "message",
       duration = 2,

     )
   })
 }
 shinyApp(ui, server)
}
}
