% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{show_tabler_toast}
\alias{show_tabler_toast}
\title{Show a tabler toast on the client}
\usage{
show_tabler_toast(id, options = NULL, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{Toast id.}

\item{options}{Toast options: see \url{https://getbootstrap.com/docs/4.3/components/toasts/}.}

\item{session}{Shiny session}
}
\description{
Show a tabler toast on the client
}
\examples{
if (interactive()) {
 library(shiny)
 ui <- tabler_page(
  tabler_toast(
    id = "toast",
    title = "Hello",
    subtitle = "now",
    "Toast body",
    img = "https://preview-dev.tabler.io/static/logo.svg"
  ),
  tabler_button("launch", "Go!", width = "25\%")
 )

 server <- function(input, output, session) {
   observe(print(input$toast))
   observeEvent(input$launch, {
     removeNotification("notif")
     show_tabler_toast(
       "toast",
       options = list(
         animation = FALSE,
         delay = 3000
       )
     )
   })

   observeEvent(input$toast, {
     showNotification(
       id = "notif",
       "Toast was closed",
       type = "warning",
       duration = 1,

     )
   })
 }

 shinyApp(ui, server)
}
}
