% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_body}
\alias{tabler_body}
\title{Create the Tabler body wrapper}
\usage{
tabler_body(..., footer = NULL)
}
\arguments{
\item{...}{Any Tabler element, especially \link{tabler_tab_items} and
\link{tabler_tab_item} if used in combination with \link{tabler_navbar}.}

\item{footer}{Slot for \link{tabler_footer}.}
}
\value{
An HTML tag containing the page elements.
}
\description{
Create the Tabler body wrapper
}
\examples{
if (interactive()) {
 library(shiny)

 ui <- tabler_page(tabler_body(h1("Hello World")))
 server <- function(input, output) {}
 shinyApp(ui, server)
}
}
