% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_card}
\alias{tabler_card}
\title{Create a Tabler card element}
\usage{
tabler_card(..., title = NULL, status = NULL, width = 6, padding = NULL)
}
\arguments{
\item{...}{Card content.}

\item{title}{Card title.}

\item{status}{Card status color. Valid statuses are given at
\url{https://preview-dev.tabler.io/docs/colors.html}.}

\item{width}{Card width. Numeric between 1 and 12 according to the
Bootstrap 4 grid system.}

\item{padding}{Card padding. Leave NULL or "sm", "md" and "lg".}
}
\value{
A card tag.
}
\description{
This is to be inserted in \link{tabler_body}.
}
\examples{
if (interactive()) {
 library(shiny)
 ui <- tabler_page(
   tabler_body(
     tabler_row(
       lapply(
        1:2,
        tabler_card,
        title = "Card",
        status = "success",
        "My amazing card"
       )
     )
   )
 )
 server <- function(input, output) {}
 shinyApp(ui, server)
}
}
