% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_dropdown}
\alias{tabler_dropdown}
\title{Create a tabler dropdown menu}
\usage{
tabler_dropdown(..., id = NULL, title, subtitle = NULL, img = NULL)
}
\arguments{
\item{...}{Slot for \link{tabler_dropdown_item}}

\item{id}{Optional: this is to be able to toggle the dropdown based on
another event like after a click on an action button.}

\item{title}{Dropdown title}

\item{subtitle}{Dropdown subtitle}

\item{img}{Dropdown image.}
}
\value{
A dropdown menu container.
}
\description{
This must be inserted in \link{tabler_navbar}
}
\seealso{
\link{show_tabler_dropdown}.
}
