% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_dropdown_item}
\alias{tabler_dropdown_item}
\title{Create Tabler dropdown item}
\usage{
tabler_dropdown_item(..., id = NULL)
}
\arguments{
\item{...}{Content.}

\item{id}{Optional. If provided, the current item will behave like
a \link{tabler_button}.}
}
\description{
Create Tabler dropdown item
}
