% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_footer}
\alias{tabler_footer}
\title{Create the Tabler footer wrapper}
\usage{
tabler_footer(left = NULL, right = NULL)
}
\arguments{
\item{left}{Left content.}

\item{right}{Right content.}
}
\value{
An HTML tag containing the footer elements.
}
\description{
Create the Tabler footer wrapper
}
\examples{
if (interactive()) {
 library(shiny)
 ui <- tabler_page(
  tabler_body(
    p("Hello World"),
    footer = tabler_footer(
      left = "Rstats, 2020",
      right = a(href = "https://www.google.com", "More")
    )
  )
 )
 server <- function(input, output) {}
 shinyApp(ui, server)
}
}
