% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_navbar}
\alias{tabler_navbar}
\title{Create the Tabler navbar}
\usage{
tabler_navbar(
  ...,
  brand_url = NULL,
  brand_image = NULL,
  nav_menu,
  nav_right = NULL
)
}
\arguments{
\item{...}{Extra elements.}

\item{brand_url}{Navbar brand url.}

\item{brand_image}{Navbar brand image.}

\item{nav_menu}{Slot for \link{tabler_navbar_menu}.}

\item{nav_right}{Right elements.}
}
\value{
The navbar HTML tag, which drives the template navigation.
}
\description{
Create the Tabler navbar
}
\examples{
if (interactive()) {
 library(shiny)

 # example without JS
 ui <- tabler_page(
  tabler_navbar(
    brand_url = "https://preview-dev.tabler.io",
    brand_image = "https://preview-dev.tabler.io/static/logo.svg",
    nav_menu = tabler_navbar_menu(
      tabler_navbar_menu_item(
        text = "Tab 1",
        icon = NULL,
        tabName = "tab1",
        selected = TRUE
      ),
      tabler_navbar_menu_item(
        text = "Tab 2",
        icon = NULL,
        tabName = "tab2"
      )
    )
  ),
  tabler_body(
    tabler_tab_items(
      tabler_tab_item(
        tabName = "tab1",
        p("Hello World")
      ),
      tabler_tab_item(
        tabName = "tab2",
        p("Second Tab")
      )
    ),
    footer = tabler_footer(
      left = "Rstats, 2020",
      right = a(href = "https://www.google.com")
    )
  )
 )
 server <- function(input, output) {}
 shinyApp(ui, server)

 # example with custom JS code to activate tabs
 shinyAppDir(system.file("chapter12/tabler_tabs", package = "OSUICode"))
}
}
