% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_navbar_menu_item}
\alias{tabler_navbar_menu_item}
\title{Create a tabler navbar menu item}
\usage{
tabler_navbar_menu_item(text, tabName, icon = NULL, selected = FALSE)
}
\arguments{
\item{text}{Item title.}

\item{tabName}{Unique tab name. Must exactly match with the
tabName parameter of \link{tabler_tab_item} on the body side.}

\item{icon}{Item icon.}

\item{selected}{Whether to select the current item at start.}
}
\value{
A Tabler navigation item tag.
}
\description{
This item is used to navigate. Must match with the
\link{tabler_tab_item} function.
}
