% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_page}
\alias{tabler_page}
\title{Create a Tabler HTML page}
\usage{
tabler_page(..., dark = TRUE, title = NULL, favicon = NULL)
}
\arguments{
\item{...}{Slot for \link{tabler_body} and \link{tabler_navbar}.}

\item{dark}{Whether to apply the dark theme. Default to TRUE.}

\item{title}{Page title.}

\item{favicon}{Site favicon.}
}
\value{
The Tabler page main wrapper.
}
\description{
This is the main wrapper
}
