% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_progress}
\alias{tabler_progress}
\title{Create a Tabler progress bar}
\usage{
tabler_progress(id = NULL, value)
}
\arguments{
\item{id}{Progress unique id.}

\item{value}{Progress value. Numeric between 0 and 100.}
}
\value{
A progress bar tag.
}
\description{
The progress bar may be updated server side. See \link{update_tabler_progress}.
}
\seealso{
\link{update_tabler_progress}
}
