% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_ribbon}
\alias{tabler_ribbon}
\title{Create a Tabler ribbon component}
\usage{
tabler_ribbon(..., position = NULL, color = NULL, bookmark = FALSE)
}
\arguments{
\item{...}{Ribbon text.}

\item{position}{Ribbon position.}

\item{color}{Ribbon color.}

\item{bookmark}{Ribbon style. FALSE by default.}
}
\value{
A ribbon shiny tag.
}
\description{
This is a good complement to \link{tabler_card}. The ribbon
parameters are listed at \url{https://preview-dev.tabler.io/docs/ribbons.html}.
}
\examples{
if (interactive()) {
 library(shiny)
 my_card <- tabler_card(title = "Ribbon", status = "info")
 my_card$children[[1]] <- my_card$children[[1]] \%>\%
 tagAppendChild(
   tabler_ribbon(
     icon("info-circle", class = "fa-lg"),
     bookmark = TRUE,
     color = "red"
   )
 )

 ui <- tabler_page(
  tabler_body(
    my_card
  )
 )
 server <- function(input, output) {}
 shinyApp(ui, server)
}
}
