% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_row}
\alias{tabler_row}
\title{Create a row container for \link{tabler_card}}
\usage{
tabler_row(...)
}
\arguments{
\item{...}{Any Tabler element.}
}
\value{
A row tag.
}
\description{
Create a row container for \link{tabler_card}
}
