% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_switch}
\alias{tabler_switch}
\title{Custom Tabler switch input}
\usage{
tabler_switch(inputId, label, value = FALSE, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:reexports]{validateCssUnit()}}.}
}
\value{
A toggle input tag.
}
\description{
Similar to the shiny checkbox
}
\seealso{
\link{update_tabler_switch}.
}
