% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_tab_item}
\alias{tabler_tab_item}
\title{Create a Tabler body tab item}
\usage{
tabler_tab_item(tabName = NULL, ...)
}
\arguments{
\item{tabName}{Unique tab name.}

\item{...}{Tab content.}
}
\description{
This works with the navbar \link{tabler_navbar_menu_item} function
}
