% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_tab_items}
\alias{tabler_tab_items}
\title{Create a Tabler body item container}
\usage{
tabler_tab_items(...)
}
\arguments{
\item{...}{Slot for \link{tabler_tab_item}.}
}
\description{
Create a Tabler body item container
}
