% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_toast}
\alias{tabler_toast}
\title{Create a Tabler toast}
\usage{
tabler_toast(id, title = NULL, subtitle = NULL, ..., img = NULL)
}
\arguments{
\item{id}{Unique toast id.}

\item{title}{Toast title.}

\item{subtitle}{Toast subtitle.}

\item{...}{Toast content.}

\item{img}{Toast image.}
}
\value{
A toast
}
\description{
Display user feedback
}
\seealso{
\link{show_tabler_toast}
}
