% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{tagAssert}
\alias{tagAssert}
\title{Check that a tag has specific properties}
\usage{
tagAssert(tag, type = NULL, class = NULL, allowUI = TRUE)
}
\arguments{
\item{tag}{Tag to check.}

\item{type}{Expected type.}

\item{class}{Expected class.}

\item{allowUI}{?}
}
\value{
An error if conditions are not met
}
\description{
Check for type and class. Raise an error if the conditions
are not fulfilled. This function is borrowed from shinydashboard.
}
\examples{
\dontrun{
 library(shiny)
 myTag <- div(class = "bg-blue")
 tagAssert(myTag, type = "div")
 tagAssert(myTag, type = "li") # will fail
 tagAssert(myTag, class = "bg-blue")
}
}
