% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{tagMatches}
\alias{tagMatches}
\title{Check that tag has specific properties}
\usage{
tagMatches(item, ..., id = NULL, name = NULL, class = NULL)
}
\arguments{
\item{item}{Tag to validate.}

\item{...}{Any attribute to check (must be named).}

\item{id}{Expected id.}

\item{name}{Expected name.}

\item{class}{Expected class.}
}
\value{
TRUE or FALSE, depending on the test result.
}
\description{
Return TRUE if a tag object matches a specific id, and/or tag name, and/or
}
\examples{
\dontrun{
 library(shiny)
 myTag <- div(class = "bg-blue")
 tagMatches(myTag, id = "d")
 tagMatches(myTag, class = "bg-blue")
}
}
