% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMobile.R
\name{update_f7_instance}
\alias{update_f7_instance}
\title{Update any UI widget on the server}
\usage{
update_f7_instance(id, options, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Widget id.}

\item{options}{New configuration list.}

\item{session}{Shiny session object.}
}
\description{
Update any UI widget on the server
}
\examples{
if (interactive()) {
 shiny::shinyAppDir(system.file("shinyMobile-pwa", package = "OSUICode"))
}
}
