% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMobile.R
\name{update_f7_tooltip}
\alias{update_f7_tooltip}
\title{Update/toggle tooltip on server side}
\usage{
update_f7_tooltip(
  id = NULL,
  selector = NULL,
  action = c("toggle", "update"),
  text = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{Tooltip id.}

\item{selector}{Tooltip selector.}

\item{action}{Action to perform: toggle or update.}

\item{text}{New text if action is update.}

\item{session}{Shiny session object.}
}
\description{
Update/toggle tooltip on server side
}
