% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{update_tabler_navbar_example}
\alias{update_tabler_navbar_example}
\title{Run the update_tabler_tab_item example}
\usage{
update_tabler_navbar_example()
}
\value{
A shiny app
}
\description{
Run the update_tabler_tab_item example
}
