% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{update_tabler_progress}
\alias{update_tabler_progress}
\title{Update a \link{tabler_progress} on the client}
\usage{
update_tabler_progress(id, value, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Progress unique id.}

\item{value}{New value.}

\item{session}{Shiny session object.}
}
\description{
Update a \link{tabler_progress} on the client
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyWidgets)
 ui <- tabler_page(
  tabler_body(
    noUiSliderInput(
      inputId = "progress_value",
      label = "Progress value",
      min = 0,
      max = 100,
      value = 20
    ),
    tabler_progress(id = "progress1", 12)
  )
 )

 server <- function(input, output, session) {
   observeEvent(input$progress_value, {
     update_tabler_progress(
       id = "progress1",
       input$progress_value
     )
   })
 }
 shinyApp(ui, server)
}
}
