% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{update_tabler_progress2}
\alias{update_tabler_progress2}
\title{Update a \link{tabler_progress} on the client}
\usage{
update_tabler_progress2(trigger, target)
}
\arguments{
\item{trigger}{Trigger id. Uses document.getElementById so
no need to add a '#'.}

\item{target}{Target id. Uses document.getElementById so
no need to add a '#'.}
}
\description{
Contrary to \link{update_tabler_progress}, this does not use session$sendCustomMessage. Everything
happens on the client. This code is invoked in the R UI.
}
