% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{update_tabler_progress_example}
\alias{update_tabler_progress_example}
\title{Run the update_tabler_progress example}
\usage{
update_tabler_progress_example()
}
\value{
A shiny app
}
\description{
Run the update_tabler_progress example
}
