% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{update_tabler_switch}
\alias{update_tabler_switch}
\title{Update \link{tabler_switch} on the client}
\usage{
update_tabler_switch(session, inputId, label = NULL, value = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}
}
\description{
Update \link{tabler_switch} on the client
}
\examples{
if (interactive()) {
 library(shiny)
 ui <- tabler_page(
  tabler_body(
    tabler_row(
      tabler_button("update", "Go!", width = "25\%", class = "mr-2"),
      tabler_switch("toggle", "Switch", value = TRUE, width = "25\%")
    )
  )
 )

 server <- function(input, output, session) {
   observe(print(input$toggle))
   observeEvent(input$update, {
     update_tabler_switch(
       session,
       "toggle",
       value = !input$toggle
     )
   })
 }

 shinyApp(ui, server)
}
}
