% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2.R
\name{updateBox}
\alias{updateBox}
\title{Collapse a \link{box2} tag.}
\usage{
updateBox(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{Box to toggle.}

\item{session}{Shiny session object.}
}
\description{
Collapse a \link{box2} tag.
}
