% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2.R
\name{updateBox2}
\alias{updateBox2}
\title{Collapse a \link{box2} tag.}
\usage{
updateBox2(inputId, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{Box to toggle.}

\item{session}{Shiny session object.}
}
\description{
Collapse a \link{box2} tag.
}
\examples{
if (interactive()) {
 library(shiny)
 library(shinyWidgets)

 ui <- fluidPage(
  # import shinydashboard deps without the need of the dashboard
  # template
  useShinydashboard(),

  tags$style("body { background-color: ghostwhite};"),

  br(),
  box2(
   title = textOutput("box_state"),
   "Box body",
   inputId = "mybox",
   collapsible = TRUE,
   plotOutput("plot")
  ),
  actionButton("toggle_box", "Toggle Box", class = "bg-success")
 )

 server <- function(input, output, session) {
  output$plot <- renderPlot({
    req(!input$mybox$collapsed)
    plot(rnorm(200))
  })

  output$box_state <- renderText({
    state <- if (input$mybox$collapsed) "collapsed" else "uncollapsed"
    paste("My box is", state)
  })

  observeEvent(input$toggle_box, {
    updateBox2("mybox")
  })

 }

 shinyApp(ui, server)

}
}
