% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{updateCustomTextInputExample}
\alias{updateCustomTextInputExample}
\title{Wrapper for Shiny App example}
\usage{
updateCustomTextInputExample(binding_step)
}
\arguments{
\item{binding_step}{Input binding step. See \link{customTextInput}.}
}
\value{
A Shiny App example
}
\description{
Wrapper for Shiny App example
}
