% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_padding}
\alias{validate_padding}
\title{Validation function}
\usage{
validate_padding(padding)
}
\arguments{
\item{padding}{Padding value to validate.}
}
\value{
An error is raised if the input padding is not valid.
}
\description{
Validate a Tabler card padding. See \link{tabler_card}
for more informations.
}
\examples{
\dontrun{
 validate_padding("xs")
 validate_padding("sm")
}
}
