% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_status}
\alias{validate_status}
\title{Validation function}
\usage{
validate_status(status)
}
\arguments{
\item{status}{Color to validate.}
}
\value{
TRUE if the test pass.
}
\description{
Validate the status of a Tabler element like \link{tabler_card}.
}
\examples{
\dontrun{
 validate_status("danger")
 validate_status("maroon")
}
}
