% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_tabName}
\alias{validate_tabName}
\title{Validate a tab name}
\usage{
validate_tabName(tabName)
}
\arguments{
\item{tabName}{Tab name to validate}
}
\value{
An error if conditions are not met.
}
\description{
Check if the provided tab name follows jQuery best practices
}
\examples{
\dontrun{
 validate_tabName("plop")
 validate_tabName("plop*+?") # fails
}
}
