% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_width}
\alias{validate_width}
\title{Validation function}
\usage{
validate_width(width)
}
\arguments{
\item{width}{Value to validate.}
}
\value{
An error is raised if conditions are not met
}
\description{
Validate a Tabler element width like \link{tabler_card}.
}
\examples{
\dontrun{
 validate_width(-1)
 validate_width(13)
 validate_width("string")
}
}
