% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket.R
\name{websocket_server}
\alias{websocket_server}
\alias{websocket_client}
\title{Initialize a websocket server}
\usage{
websocket_server(host = "127.0.0.1", port = 8080)

websocket_client(host = "127.0.0.1", port = 8080)
}
\arguments{
\item{host}{A string that is a valid IPv4 address that is owned by this server,
or "0.0.0.0" to listen on all IP addresses.}

\item{port}{A number or integer that indicates the server port that should be listened on.
Note that on most Unix-like systems including Linux and Mac OS X, port numbers smaller than
1025 require root privileges.}
}
\value{
A client websocket connection
}
\description{
This has to be used with \link{websocket_client}.
}
\examples{
library(OSUICode)
server <- websocket_server()
client_1 <- websocket_client()
client_2 <- websocket_client()
client_1$send("Hello from client 1")
client_2$send("Hello from client 2")
client_1$close()
client_2$send("Only client 2 is here")
Sys.sleep(1)
server$stop()
}
