% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMessageItem.R
\name{dropdownMenu}
\alias{dropdownMenu}
\title{Create a Boostrap 4 dashboard dropdown menu}
\usage{
dropdownMenu(
  ...,
  type = c("messages", "notifications", "tasks"),
  badgeStatus = "primary",
  icon = NULL,
  headerText = NULL,
  .list = NULL,
  href = NULL
)
}
\arguments{
\item{...}{Items to put in the menu. Typically, message menus should contain
\code{\link[bs4Dash]{messageItem}}s, notification menus should contain
\code{\link[bs4Dash]{notificationItem}}s, and task menus should contain
\code{\link[bs4Dash]{taskItem}}s.}

\item{type}{The type of menu. Should be one of "messages", "notifications",
"tasks".}

\item{badgeStatus}{The status of the badge which displays the number of items
  in the menu. This determines the badge's color. Valid statuses are defined as follows:
\itemize{
  \item \code{primary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#007bff")}.
  \item \code{secondary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6c757d")}.
  \item \code{info}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#17a2b8")}.
  \item \code{success}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#28a745")}.
  \item \code{warning}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ffc107")}.
  \item \code{danger}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#dc3545")}.
}.
A value of \code{NULL} means to not display a badge.}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{headerText}{An optional text argument used for the header of the
dropdown menu (this is only visible when the menu is expanded). If none is
provided by the user, the default is "You have \code{x} messages," where
\code{x} is the number of items in the menu (if the \code{type} is
specified to be "notifications" or "tasks," the default text shows "You
have \code{x} notifications" or  "You have \code{x} tasks," respectively).}

\item{.list}{An optional list containing items to put in the menu Same as the
\code{...} arguments, but in list format. This can be useful when working
with programmatically generated items.}

\item{href}{An optional URL to link to.}
}
\description{
Build an adminLTE3 dashboard dropdown menu
}
