% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_example}
\alias{get_example}
\title{Run Shiny app example}
\usage{
get_example(path)
}
\arguments{
\item{path}{App location.}
}
\description{
Run Shiny app example
}
\examples{
if (interactive()) {
 get_example("dj-system")
}
}
