% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_example}
\alias{run_example}
\title{Wrapper for running shiny app example}
\usage{
run_example(path)
}
\arguments{
\item{path}{App location.}
}
\description{
Similar to shiny::shinyAppDir
}
\examples{
if (interactive()) {
 run_example("dj-system")
}
}
