% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabler.R
\name{tabler_navbar_menu}
\alias{tabler_navbar_menu}
\title{Create the Tabler navbar menu}
\usage{
tabler_navbar_menu(...)
}
\arguments{
\item{...}{Slot for \link{tabler_navbar_menu_item}.}
}
\description{
Create the Tabler navbar menu
}
