% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2.R
\name{updateBox2}
\alias{updateBox2}
\title{Update a \link{box2} tag.}
\usage{
updateBox2(
  id,
  action = c("toggle", "update"),
  options = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{id}{If passed, the item will behave like an action button.}

\item{action}{Action to trigger: either collapse, remove, restore or update.}

\item{options}{If action is update, a list of new options to configure the box, such as
\code{list(title = "new title", status = NULL, solidHeader = FALSE, 
background = "red", width = 6, height = "200px", collapsible = FALSE, closable = FALSE)}.
If the box had a background/status (any item that may be NULL), you must explicitly pass background = NULL, 
if you want to remove the background value.}

\item{session}{Shiny session object.}
}
\description{
Update a \link{box2} tag.
}
